Fix the read/write files implementation
I was using fgets to read the files, but I learned that this way you can
accidentally truncate files because the function stops after an EOF or a
newline.

Using the command cor and cat I saw that the command cat preserved the
content inside a binary file and with the command cor the binary file
had truncated lines.

So I needed to use fread as it reads raw data and fwrite when printing
files with NO_COLOR set (normal cat behavior).