Refactor code and change options and exit status
Added --help and --version options and updated the help information.

Added the functionality of passing multiple single characters options in
the same argument. Example: '-cn'.

When the FILE *fp is NULL, set status to 1 instead of bitwise OR
assignment with errno.

Added a debug target in the Makefile.

Added config.mk for setting the flags and variables